Refactor: #256 elderId 타입을 Int에서 Long으로 통일
- Service, DTO, Repository, RepositoryImpl 전반 Int → Long 변경
- Navigation Route, ViewModel, Screen의 elderId 파라미터 타입 통일
- ElderIds, ElderHealthInfo, ElderSubscription 등 모델 타입 일괄 변경

+++ app/src/main/java/com/konkuk/medicarecall/data/api/elders/HealthService.kt
@@ -9,7 +9,7 @@ import de.jensklingenberg.ktorfit.http.Query
 interface HealthService {
     @GET("elders/{elderId}/health-analysis")
     suspend fun getDailyHealth(
-        @Path("elderId") elderId: Int,
+        @Path("elderId") elderId: Long,
         @Query("date") date: String,
     ): Response<HealthResponseDto>
 }
