+++ app/src/main/java/com/konkuk/medicarecall/data/repository/GlucoseRepository.kt
@@ -3,5 +3,5 @@ package com.konkuk.medicarecall.data.repository
 import com.konkuk.medicarecall.data.dto.response.GlucoseResponseDto
 
 interface GlucoseRepository {
-    suspend fun getGlucoseGraph(elderId: Int, counter: Int, type: String): Result<GlucoseResponseDto>
+    suspend fun getGlucoseGraph(elderId: Long, counter: Int, type: String): Result<GlucoseResponseDto>
 }
