+++ app/src/main/java/com/konkuk/medicarecall/data/repository/MealRepository.kt
@@ -4,5 +4,5 @@ import com.konkuk.medicarecall.ui.feature.homedetail.meal.viewmodel.MealUiState
 import java.time.LocalDate
 
 interface MealRepository {
-    suspend fun getMealUiStateList(elderId: Int, date: LocalDate): List<MealUiState>
+    suspend fun getMealUiStateList(elderId: Long, date: LocalDate): List<MealUiState>
 }
