+++ app/src/main/java/com/konkuk/medicarecall/data/repository/SleepRepository.kt
@@ -4,5 +4,5 @@ import com.konkuk.medicarecall.domain.model.Sleep
 import java.time.LocalDate
 
 interface SleepRepository {
-    suspend fun getSleepData(elderId: Int, date: LocalDate): Result<Sleep>
+    suspend fun getSleepData(elderId: Long, date: LocalDate): Result<Sleep>
 }
